requirements.txt
uvicorn[standard]==0.22.0
fastapi==0.103.1
pandas==2.1.0
flake8==6.0.0
pre-commit==3.0.2
pytest==7.1.3
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.1
pyarrow==15.0.0
